Make CoordinatesModule class a partial one
